// src/components/Media/types.ts
import type { StaticImageData } from 'next/image'
import type { ElementType, Ref } from 'react'

import type { Media as MediaType } from '@/payload-types'
export interface Props {
	alt?: string
	className?: string
	fill?: boolean // for NextImage only
	htmlElement?: ElementType | null
	imgClassName?: string
	onClick?: () => void
	onLoad?: () => void
	priority?: boolean // for NextImage only
	ref?: Ref<HTMLImageElement | HTMLVideoElement | null>
	layout?: 'fill' | 'responsive' | 'fixed' | 'intrinsic'
	objectFit?: 'cover' | 'contain' | 'fill' | 'none' | 'scale-down'
	resource?: MediaType | string | number // for Payload media
	size?: string // for NextImage only
	src?: StaticImageData // for static media
	videoClassName?: string
	blurDataURL?: string
	width?: number
	height?: number
	isBlurred?: boolean
}